Removed and renamed files.
